SwiftUI
Putting together a plan for Learning how to code in a full stack using SwiftUI, C++, Java, JavaScript,JSON by taking a break on coding on the computer to read books and to create a schedule for coding, practice, reading and resource finding.

All future commits will hold only the set of goals with a plan deadline for execution or updated deadlines for how things will work.

Goal is to be SwiftUI ready in 18 weeks.